requirements.txt
ipyleaflet
ipywidgets
matplotlib
numpy
opencv-python
pandas
psycopg2-binary
psycopg2
pymysql
python-dotenv
requests
scikit-learn
seaborn
sqlalchemy
sympy
xgboost
openpyxl